% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{authenticate}
\alias{authenticate}
\title{authenticate}
\usage{
authenticate(axapi.uri = "https://ui.anametrix.com/api", username, password)
}
\arguments{
\item{axapi.uri}{default is "https://ui.anametrix.com/api". URI is different for R Jobs.}

\item{username}{most often an anametrix email (e.g. "demo.user@demoanametrix.com")}

\item{password}{associated password}
}
\description{
Authenticates to the Anametrix platform.
}
\details{
return a list of authentication elements (API URL, access token) used to download and upload data to the platform.
}
\examples{
username <- "demo.user@demoanametrix.com"
password <- "password"
auth     <- authenticate("https://demo.anametrix.com/api", username, password)

}
\seealso{
Other authentication functions: \code{\link{authenticationCheck}}
}

