% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{authenticationCheck}
\alias{authenticationCheck}
\title{authenticationCheck}
\usage{
authenticationCheck(auth)
}
\arguments{
\item{auth}{Authentication instance. result of \code{\link{authenticate}} function}
}
\value{
boolean.
}
\description{
checking if well authenticated
}
\details{
Checks if an existing authentication instance is still valid to access the platform.
}
\examples{
username <- "demo.user@demoanametrix.com"
password <- "password"
auth     <- authenticate("https://demo.anametrix.com/api", username, password)
authenticationCheck(auth)

}
\seealso{
Other authentication functions: \code{\link{authenticate}}
}

