% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{getAxDatasetFromEncodedQuery}
\alias{getAxDatasetFromEncodedQuery}
\title{getAxDatasetFromEncodedQuery}
\usage{
getAxDatasetFromEncodedQuery(auth, encoded.query, start.date, end.date, verbose)
}
\arguments{
\item{auth}{Authentication instance. result of \code{\link{authenticate}} function}

\item{encoded.query}{Base64-encoded query XML corresponding to Anametrix API requirements.
Contact your product manager for assistance.}

\item{start.date}{Specifies start date of the query (character string, format: YYYY-MM-DD).}

\item{end.date}{Specifies end date of the query (character string, format: YYYY-MM-DD).}

\item{verbose}{Specifies whether to print out notifications. Example: FALSE}
}
\value{
R data frame containing the observations requested from the Anametrix API table
}
\description{
Extracts data from Anametrix API using Base64-encoded Anametrix query, and returns R data frame
}
\seealso{
Other dowload/upload functions: \code{\link{getAnametrixDataset}},
  \code{\link{getTableConfiguration}},
  \code{\link{uploadData}}
}

