% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{printTable}
\alias{printTable}
\title{printTable}
\usage{
printTable(table.object)
}
\arguments{
\item{table.object}{Configuration of a table. Use  \code{\link{getTableConfiguration}} to get it}
}
\description{
Prints configuration of a table object extracted from Anametrix API
}

