% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{uploadData}
\alias{uploadData}
\title{uploadData}
\usage{
uploadData(auth, report.suite.id, table.object, dataframe, columns, tag,
  truncation = "none")
}
\arguments{
\item{auth}{Authentication instance. See \code{\link{authenticate}}.}

\item{report.suite.id}{ID number (numeric or character) of the data suite.}

\item{table.object}{See \code{\link{getTableConfiguration}} to get it}

\item{dataframe}{Data to upload. Must have the same number of columns as the table.object}

\item{columns}{Array of column names that need to be uploaded. Example: c("city","country","male_population")
Use printTable(tableObject) to see all column names.}

\item{tag}{tags are used to identify the origin of the upload. Tags are associated to a user, therefore data associated with another user's tag can't be removed or modified.}

\item{truncation}{"none" or "tag". If "none", new data is just bound to the existing table. If "tag", data associated with the same tag is overwritten.}
}
\description{
uploads data from R data.frame into Anametrix table
}
\seealso{
Other dowload/upload functions: \code{\link{getAnametrixDataset}},
  \code{\link{getAxDatasetFromEncodedQuery}},
  \code{\link{getTableConfiguration}}
}

