\name{anametrix-package}
\alias{anametrix-package}
\alias{anametrix}
\docType{package}
\title{
Bidirectional connector to Anametrix API
}
\description{
Provides function(s) to extract data from Anametrix API into R data structure(s).
Provides function(s) to upload data from R into Anametrix table(s).
}
\details{
\tabular{ll}{
Package: \tab anametrix\cr
Type: \tab Package\cr
Version: \tab 1.5\cr
Date: \tab 2012-12-13\cr
License: \tab LGPL\cr
}
In R install package: install.package("anametrix"). Load package: library("anametrix")
}
\author{
Author, Roman Jugai

Maintainer: Roman Jugai <roman@anametrix.com>
}
\references{
http://www.anametrix.com
}
\keyword{connection}
\keyword{database}
