\name{anametrix-package}
\alias{anametrix-package}
\alias{anametrix}
\alias{getFullTableListXML}
\alias{getTableConfiguration}
\alias{printTable}
\alias{constructQueryXML}
\alias{queryReader}
\alias{getColumnTypeId}
\alias{authenticate}
\docType{package}
\title{
Connects to Anametrix API
}
\description{
Provides function(s) to extract data from Anametrix API into R data structures.
}
\details{
\tabular{ll}{
Package: \tab anametrix\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2012-05-29\cr
License: \tab LGPL\cr
}
In R install package: install.package("anametrix"). Load package: library("anametrix")
}
\author{
Author, Roman Jugai

Maintainer: Roman Jugai <roman@anametrix.com>
}
\references{
http://www.anametrix.com
}
\keyword{connection}
\keyword{database}
