\name{printTable}
\alias{printTable}
\title{
Prints Table object configuration
}
\description{
Prints configuration of a table object extracted from Anametrix API
}
\usage{
printTable(tableObject)
}
\arguments{
 \item{tableObject}{
Configuration of a table. Use getTableConfiguration(auth, reportsuiteId, tableName) to get it.
}
}
\author{
Author, Roman Jugai
}

\keyword{connection}
\keyword{database}